requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.3
django-cbv==0.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
ecdsa==0.17.0
fastapi==0.78.0
fastapi-jwt-auth==0.5.0
fastapi-pagination==0.9.3
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.1.2
jwt-auth==0.4.3
MarkupSafe==2.1.1
mysqlclient==2.1.0
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==1.7.1
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
sqlalchemy2-stubs==0.0.2a23
sqlmodel==0.0.6
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.17.6
